services/django/requirements.txt
Django==5.0.7
mysqlclient==2.2.4
django-redis==5.4.0
redis==5.0.8
boto3==1.35.0
django-storages==1.14.4
djangorestframework==3.15.2
PyJWT==2.9.0
python-dotenv==1.0.1
gunicorn==22.0.0
whitenoise==6.7.0
requests==2.32.3
django-cors-headers==4.4.0

